featured products now show price range and image of product. Products can only be featured if in stock.

+++ src/main/java/com/brokeshirts/ecom/models/Inventory.java
@@ -54,9 +54,7 @@ public class Inventory {
         this.sku = sku;
     }
 
-    public double getPrice() {
-        return price;
-    }
+    public Float getPrice() { return (float) price; }
 
     public void setPrice(double price) {
         this.price = price;
